Don't touch SDK dir when installing.
It is becoming more likely that people will have the Android SDK installed
by their distros package manager. In such a case, overwriting and adding files
in the Android SDK dir might be problematic. I've tried to solve this problem
by prepending the adb-wrapper dir to the PATH variable and symlinking
`adb_wrapper.sh` to `adb`. This way, our `adb` is preferred over the SDK
version.

The trade-offs are:
 * Tools which don't use $PATH will not get to use adb-wrapper (might be good
for things like debuggers, but usually is annoying).
 * The install process is a little bit more clunky. I'd hope that people working
   with the Android SDK have enough technical know-how to follow instructions
   to edit their ~/.bashrc file though.